split out message posting/editing/deleting
